feat: support unison (#41)
* feat: support unison using voices

* refactor: create function to optimize output level by unison

* fix: incorrect variable name

+++ include/synth2/helper.h
@@ -18,6 +18,6 @@
 #include <stdint.h>
 
 /// Convert midi1 key number (0..127) into frequency.
-double k2f(int16_t key, int16_t cent);
+double k2f(int16_t key, double cent);
 
 #endif  // SYNTH2_HELPER_H_
